scripts/validated-pattern-converter/pyproject.toml
[tool.poetry]
name = "validated-pattern-converter"
version = "2.0.0"
description = "A tool to convert OpenShift/Kubernetes projects into Red Hat Validated Patterns"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "Apache-2.0"
repository = "https://github.com/validatedpatterns/pattern-converter"
keywords = ["openshift", "kubernetes", "validated-patterns", "gitops", "helm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]
packages = [{include = "vpconverter"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
pyyaml = "^6.0.1"
gitpython = "^3.1.40"
jinja2 = "^3.1.2"
colorama = "^0.4.6"
rich = "^13.7.0"
pydantic = "^2.5.0"
typing-extensions = "^4.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
isort = "^5.13.0"
pre-commit = "^3.6.0"
pytest-mock = "^3.12.0"
types-pyyaml = "^6.0.12.12"

[tool.poetry.scripts]
vp-convert = "vpconverter.cli:main"
validated-pattern-converter = "vpconverter.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --cov=vpconverter --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"